+++ goaltimer.cpp
@@ -240,7 +240,7 @@ int16_t CGoalTimer::EditNew(									// Returns 0 if successfull, non-zero other
 			RGuiItem* pgui = plb->AddString((char*) m_rstrNetName);
 			pgui->m_lId = GetInstanceID();
 			pgui->m_bcUser = NavNetListPressedCall;
-			pgui->m_ulUserInstance = (unsigned long) this;
+			pgui->m_ulUserInstance = (uint32_t) this;
 			plb->AdjustContents();
 			plb->SetSel(pgui);
 		}
